MRSI

Documentation for MRSI.

MRSI.fourier_transform

MRSI.get_dft_matrix

MRSI.initialize_header_storage

MRSI.reconstruct

MRSI.fourier_transform — Method

fourier_transform(kdata, kspace_coordinates, n_grid)

Fourier transform from kspace to a slice. Can be one circle or whole slice

MRSI.get_dft_matrix — Method

get_dft_matrix(circle::Circle, n_grid)

Retrieves the DFT matrix for a given circle. Memoizes the matrix for the same circle

MRSI.initialize_header_storage — Method

headers = initialize_header_storage(info)

Returs empty Circle array of size [nPart][nCircles] Each Circle can store all headers corresponding to that part and circle.

MRSI.reconstruct — Method

csi = reconstruct(filename; combine=:auto, datatype=ComplexF32, ice=false, old_headers=false, mmap=true, lipid_decon=false, lipid_mask, brain_mask, do_fov_shift=true, do_freq_cor=true, do_dens_comp=true, conj_in_beginning=true)

Reconstructs a SIEMENS dat file

Coil combine is performed automatically for AC datasets. Set combine to false|true to force coil combination (including normalization by the patrefscan). Set ice to true to use calculated radii for density compensation instead of reading them from the headers. Set old_headers to true for older dat files with part and circle stored in LIN Set mmap to false to compute in RAM or a filename or folder for storing the temporary result array. Set lipid_decon to :L1 or :L2 to perform lipid decontamination. Set gradient_delay to [[x1,y1], [x2,y2], [x3,y3]] to apply a gradient delay correction For lipid decontamination, pass the filenames of Float32 raw files via lipid_mask and brain_mask. Alternatively, use lipid_mask=:from_spectrum. If nothing is provided, a full mask is used. Set settings for lipid decontamination: L2_beta=0.1f0, L1_n_loops=5

csi, info = reconstruct(filename, type; options...)

Reconstruction without coil combination. type can be :ONLINE or :PATREFSCAN info is a Dict containing scan information.